bugfix:文件上传功能bugfix

+++ app/src/main/java/com/jamgu/hwstatistics/net/upload/DataUploader.kt
@@ -1,11 +1,15 @@
 package com.jamgu.hwstatistics.net.upload
 
 import android.content.Context
+import com.jamgu.common.thread.ThreadPool
 import com.jamgu.common.util.log.JLog
 import com.jamgu.common.util.preference.PreferenceUtil
 import com.jamgu.hwstatistics.net.Network
 import com.jamgu.hwstatistics.net.RspModel
 import com.jamgu.hwstatistics.power.mobiledata.network.NetWorkManager
+import com.jamgu.hwstatistics.util.TimeExtensions
+import com.jamgu.hwstatistics.util.getCurrentDateString
+import com.jamgu.hwstatistics.util.timeMillsBetween
 import io.reactivex.Observer
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
@@ -27,10 +31,11 @@ object DataUploader {
     const val USER_PREFIX = ""
 
     private fun upload(context: Context, file: File) {
+        JLog.d("upload", "upload file = ${file.path}")
         val isScreenOff = PreferenceUtil.getCachePreference(context, 0).getBoolean((DataSaver.TAG_SCREEN_OFF), false)
         if (!isScreenOff) {
-            JLog.d(TAG, "uploading when screen off.")
-            DataSaver.addTestTracker(context, "uploading when screen off, file = ${file.absolutePath}")
+            JLog.d(TAG, "uploading when screen off, file = ${file.path}")
+            DataSaver.addDebugTracker(context, "uploading when screen off, file = ${file.absolutePath}")
             return
         }
         try {
@@ -46,7 +51,7 @@ object DataUploader {
                 .build()
 
             if (!isNetWorkEnable(context)) {
-                DataSaver.addErrorTracker(context, "network error, upload failed, file = $filePath")
+                DataSaver.addInfoTracker(context, "network error, upload failed, file = $filePath")
                 return
             }
 
@@ -56,20 +61,20 @@ object DataUploader {
                 .subscribe(object: Observer<RspModel> {
                     override fun onNext(rspModel: RspModel) {
                         if (rspModel.getCode() == 0) {
-                            if (!checkIfNeedDelete(file)) {
-                                // 文件未成功删去
-                                DataSaver.addErrorTracker(context, "delete file failed, filepath = $filePath")
+                            if (!checkIfNeedDeleteAfterUpload(file)) {
+                                // 文件不需要删去
+                                DataSaver.addInfoTracker(context, "does not need to be deleted, filepath = $filePath")
                             } else {
                                 // 看看它的父目录还是否有文件，没有的话，删掉父目录
                                 val dirFile = file.parentFile ?: return
-                                if (dirFile.isDirectory && dirFile.listFiles().isNullOrEmpty()) {
+                                if (dirFile.isDirectory && dirFile.list().isNullOrEmpty()) {
                                     if (!dirFile.delete()) {
-                                        DataSaver.addErrorTracker(context, "parent file delete failed, filepath = ${dirFile.absolutePath}")
+                                        DataSaver.addInfoTracker(context, "parent file delete failed, filepath = ${dirFile.absolutePath}")
                                     }
                                 }
                             }
                         } else {
-                            DataSaver.addErrorTracker(context, "upload file failed, code = ${rspModel.getCode()}, " +
+                            DataSaver.addInfoTracker(context, "upload file failed, code = ${rspModel.getCode()}, " +
                                     "msg = ${rspModel.getMsg()}, filepath = $filePath")
                             JLog.d(TAG, "upload file failed, code = ${rspModel.getCode()}, msg = ${rspModel.getMsg()}, filepath = $filePath")
                         }
@@ -79,27 +84,85 @@ object DataUploader {
                     }
 
                     override fun onError(e: Throwable) {
-                        JLog.d(TAG, "onError")
-                        DataSaver.addErrorTracker(context, "filepath = ${file.absolutePath}, error when uploading, e = ${e.stackTrace}")
+                        JLog.d(TAG, "filepath = ${file.absolutePath}, error when uploading, e = ${e.stackTrace}")
+                        DataSaver.addInfoTracker(context, "filepath = ${file.absolutePath}, error when uploading, e = ${e.stackTrace}")
                     }
 
                     override fun onComplete() {
                     }
                 })
         } catch (e: Exception) {
-            DataSaver.addErrorTracker(context, "filepath = ${file.absolutePath}, error happened when uploading")
+            DataSaver.addInfoTracker(context, "filepath = ${file.absolutePath}, error happened when uploading")
         }
     }
 
     /**
      * @return 判断该文件是否满足删去的规则，是的话返回该文件是否成功删除
      */
-    private fun checkIfNeedDelete(file: File): Boolean {
+    private fun checkIfNeedDeleteAfterUpload(file: File): Boolean {
         return if (file.absolutePath.contains(DataSaver.CACHE_ROOT_DIR)) {
             if (file.exists() && !file.isDirectory) {
-                file.delete()
+                // session 数据文件，保留一周，一周后清除
+                val sessionFileEndTime = getSessionFileEndTime(file)
+                if (sessionFileEndTime.isNotEmpty()) {
+                    val duration = getCurrentDateString().timeMillsBetween(sessionFileEndTime)
+                    if (duration >= TimeExtensions.ONE_WEEK) {
+                        return file.delete()
+                    }
+                }
+                false
             } else false
         } else false
+    }
+
+    private fun innerRecursivelyUpload(context: Context, file: File, timeStamp: String) {
+        if (!file.exists()) return
+
+        val childFiles = file.listFiles()
+//        JLog.d(TAG, "file ${file.path}, listFiles${childFiles == null}' size = ${childFiles?.size ?: 0}")
+        childFiles?.forEach { child ->
+            val directory = child.isDirectory
+            JLog.d(TAG, "recursivelyUpload, directory path = ${child.path}, $timeStamp")
+            if (directory) {
+//                JLog.d(TAG, "recursivelyUpload, directory path = ${child.path}, $timeStamp")
+                if (child.list()?.isEmpty() == true) {
+                    child.delete()
+                    DataSaver.addInfoTracker(context, "filepath = ${file.absolutePath}, error happened when uploading")
+                } else {
+                    innerRecursivelyUpload(context, child, timeStamp)
+                }
+            } else {
+//                JLog.d(TAG, "recursivelyUpload, child ----- path = ${child.path}, $timeStamp")
+                val sessionEndTime = getSessionFileEndTime(child)
+                val canUpload = if (sessionEndTime.isNotEmpty()) {
+                    // session文件目录是否完整，完整才能上传
+                    sessionEndTime < timeStamp
+                } else {
+                    true
+                }
+                if (canUpload) {
+                    upload(context, child)
+                }
+            }
+        }
+    }
+
+    private fun isSessionFile(file: File): Boolean {
+        val filePath = file.path ?: return false
+        return filePath.contains(DataSaver.SESSION_FILE_INFIX)
+    }
+
+    /**
+     * @return 返回session文件的截止时间，如果该文件不是session文件，返回 ""
+     */
+    private fun getSessionFileEndTime(file: File): String {
+        val filePath = file.path ?: return ""
+        return if (isSessionFile(file)) {
+            val childPathSplits = filePath.split(DataSaver.SESSION_FILE_INFIX)
+            if (childPathSplits.size == 2 && childPathSplits[1].isNotEmpty()) {
+                childPathSplits[1]
+            } else ""
+        } else ""
     }
 
     /**
@@ -107,32 +170,7 @@ object DataUploader {
      * @param timeStamp session文件，只会上传记录时间完成在timeStamp之前的文件
      */
     fun recursivelyUpload(context: Context, file: File, timeStamp: String) {
-        if (!file.exists()) return
-
-        val childFiles = file.listFiles()
-        childFiles?.forEach { child ->
-            if (child.isDirectory) {
-                if (child.listFiles().isNullOrEmpty()) {
-                    child.delete()
-                } else {
-                    val childPath = child.absolutePath ?: return
-                    // session文件目录是否完整，完整才能上传
-                    val canUpload = if (childPath.contains(DataSaver.FILE_INFIX)) {
-                        val childPathSplits = childPath.split(DataSaver.FILE_INFIX)
-                        // session 文件必须完整才能上传
-                        JLog.d(TAG, "path = $childPath, ${childPathSplits[1]}, $timeStamp")
-                        childPathSplits.size == 2 && childPathSplits[1].isNotEmpty() && childPathSplits[1] < timeStamp
-                    } else {
-                        true
-                    }
-                    if (canUpload) {
-                        recursivelyUpload(context, child, timeStamp)
-                    }
-                }
-            } else {
-                upload(context, child)
-            }
-        }
+        ThreadPool.runIOTask { innerRecursivelyUpload(context, file, timeStamp) }
     }
 
     private fun isNetWorkEnable(context: Context): Boolean {
